[Embedding] remove kernel functions from operator (#545)
* [Embedding] remove kernel functions from operator

* Added formatting

* Removed comment

* Fixed embedding kernels cpp error
