+++ src/ops/fused.cu
@@ -17,6 +17,7 @@
 #include "flexflow/model.h"
 #include "flexflow/ops/batch_norm.h"
 #include "flexflow/ops/element_unary.h"
+#include "flexflow/ops/flat.h"
 #include "flexflow/ops/embedding.h"
 #include "flexflow/ops/fused.h"
 #include "flexflow/ops/kernels/batch_matmul_kernels.h"
@@ -24,6 +25,7 @@
 #include "flexflow/ops/kernels/conv_2d_kernels.h"
 #include "flexflow/ops/kernels/dropout_kernels.h"
 #include "flexflow/ops/kernels/element_binary_kernels.h"
+#include "flexflow/ops/kernels/embedding_kernels.h"
 #include "flexflow/ops/kernels/flat_kernels.h"
 #include "flexflow/ops/kernels/linear_kernels.h"
 #include "flexflow/ops/kernels/pool_2d_kernels.h"
@@ -355,13 +357,13 @@ __host__ void FusedOp::forward_task(Task const *task,
         }
 
         assert(my_input_accessor[0].data_type == DT_INT64);
-        Embedding::forward_kernel_wrapper(m,
-                                          my_input_accessor[0],
-                                          my_output_accessor[0],
-                                          my_weight_accessor[0],
-                                          in_dim,
-                                          out_dim,
-                                          effective_batch_size);
+        Kernels::Embedding::forward_kernel_wrapper(m,
+                                                   my_input_accessor[0],
+                                                   my_output_accessor[0],
+                                                   my_weight_accessor[0],
+                                                   in_dim,
+                                                   out_dim,
+                                                   effective_batch_size);
         break;
       }
       case OP_RELU:
@@ -783,13 +785,13 @@ __host__ void FusedOp::backward_task(Task const *task,
           assert(effective_batch_size * in_dim ==
                  my_input_accessor[0].domain.get_volume());
         }
-        Embedding::backward_kernel_wrapper(m,
-                                           my_input_accessor[0],
-                                           my_output_grad_accessor[0],
-                                           my_weight_grad_accessor[0],
-                                           in_dim,
-                                           out_dim,
-                                           effective_batch_size);
+        Kernels::Embedding::backward_kernel_wrapper(m,
+                                                    my_input_accessor[0],
+                                                    my_output_grad_accessor[0],
+                                                    my_weight_grad_accessor[0],
+                                                    in_dim,
+                                                    out_dim,
+                                                    effective_batch_size);
         break;
       }
       case OP_LINEAR: {
